+++ com.unity.shadergraph/Editor/Drawing/Inspector/PropertyDrawers/GraphDataPropertyDrawer.cs
@@ -133,7 +133,7 @@ namespace UnityEditor.ShaderGraph.Drawing.Inspector.PropertyDrawers
                 newValue => { m_postChangeConcretePrecisionCallback((ConcretePrecision) newValue); },
                 graphData.concretePrecision,
                 "Precision",
-                ConcretePrecision.Float,
+                ConcretePrecision.Single,
                 out var propertyVisualElement));
 
             propertySheet.Add(GetSettings(graphData, () => this.m_postChangeTargetSettingsCallback()));